refactor : recommendation 컴포넌트 분리
